% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gim.R
\name{GIM}
\alias{GIM}
\title{GIM}
\usage{
GIM(out, full = TRUE, B, B2, cluster = NA, time = NA)
}
\arguments{
\item{out}{: Output from glm function}

\item{full}{: TRUE if you want to perform the full GIM test, FALSE if you want to
quickly check if your model is misspecified}

\item{B}{: Number of new datasets to test}

\item{B2}{: Number of bootstraps for each new dataset}

\item{cluster}{: For clustered data}

\item{time}{: For time series data}
}
\description{
Takes in glm output, performs GIM test, returns estimated coefficients,
  classic standard errors, robust standard errors, rule of thumb for
  misspecified model, GIM test statistic and p-value.
}
\examples{
# estimate model
out <- glm(Sepal.Length ~ ., data = iris)
GIM(out)

}
